episode 9

+++ app/Http/Controllers/ArticlesController.php
@@ -17,7 +17,7 @@ class ArticlesController extends Controller
     public function index()
     {
         return view('articles.index', [
-            'articles' => Article::all(),
+            'articles' => Article::visibleTo(Auth::user())->get(),
         ]);
     }
 
@@ -26,6 +26,8 @@ class ArticlesController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Article::class);
+
         return view('articles.create');
     }
 
@@ -68,6 +70,8 @@ class ArticlesController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('delete', $article);
+
         $article->delete();
 
         return redirect()->route('articles.index');
